Move print_unknown_options() to engine.py
That is better for specifying typing.
